[Backport 1.3] Minimum viable integration tests framework (#3649)
We've been authoring our new test cases in the integration test
framework that tackle difficult problems that we've been unable to
validate in the older framework around concurrency and environment
configuration. As part of getting the fix for handling unauthenticated
requests better - those tests were written in the new framework,
backporting it into 1.3 so we can pull everything in.

Signed-off-by: Peter Nied <petern@amazon.com>